+++ lib/convert.ts
@@ -1,0 +1,7 @@
+type Options = {
+  [key in string]: string
+}
+
+export default function convert(template: string, options?: Options): string {
+  return template;
+}
